+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Remove add yarn.lock on Dockerfile 
 - Update celebrate and joi package
 - Make port optional, so user can change them upon generation
+- Changed validation into validator with joi schema
 
 # Version 1.0 - 3.0.0